Make FileInfo checksum generation optional
In some environments FITS is being called on files with known
checksums, whose checksums are being checked in other parts of the
workflow. Since MD5 hash generation is fairly computationally
expensive, it's a significant amount of extra CPU time to spend on
metadata that might not be used.

This commit makes MD5 checksum generation optional, controlled by the
option.enable-checksum parameter in the fits.xml file.